fix website not showing up
react enters a rendering loop (component loads, setLoading is set to false, so react rerenders, and setLoading is set to false, and react rerenders, and so on), so we remove the extra setLoading check outside the useEffect function and make a useEffect dependency on getUserStoreSignedIn instead.

The ternary conditional in the return value is replaced with a Navigate to login instead of returning the child value, so we don't encounter a missing sidebar issue, and we get redirected to the login page whenever the sidebar state isn't synced with the login state.